+++ src/main/java/com/example/shortudy/domain/upload/dto/ShortsUploadCompleteRequest.java
@@ -1,4 +1,4 @@
-package com.example.shortudy.domain.shorts.upload.dto;
+package com.example.shortudy.domain.upload.dto;
 
 import jakarta.validation.constraints.NotBlank;
